+++ Anarchy/ChangeLog.md
@@ -1,6 +1,7 @@
 ﻿## Update V1.7.0
-* Added Optional Network Anarchy Tool options: Force Ground, Force Elevated, Force Tunnel, and Constant Slope.
+* Added Optional Network Anarchy Tool options: Force Ground, Force Elevated, Force Tunnel, Constant Slope, and Expanded Elevation Range.
 * Added Optional Network Upgrade Overhaul so network upgrades can be applied during placement or replacement. Includes retaining walls, quays, trees, grass, wide sidewalks, lighting, medians, etc.
+  Note: Replace tool mode will by default remove/add upgrades based on tool options instead of preserving upgrades when possible. Option to have the opposite behavior be default.
 * Added Anarchy Components Tool which lets you Add or remove elevation lock or anarchy components using a tool with radius or single selection.
 * Added Optional Network Upgrade assets from Extended Road Upgrades with permission from ST-Apps. Compatible with ERU if you choose not to unsubscribe.
 * Fixes to Extended Road Upgrades including: retaining wall node issues and fire and rescue info view issue.