+++ Anarchy/ChangeLog.md
@@ -1,2 +1,3 @@
 ﻿## Patch V1.7.1
-* Fix some problems with replacing elevated networks and tunnels. There may still be issues with these.
+* Fixed regression that didn't allow retaining walls and quays to be removed.
+* Improved replace tool mode's ability to swap configurations.