+++ Anarchy/ChangeLog.md
@@ -1,3 +1,5 @@
-﻿## Patch V1.7.1
+﻿## Patch V1.7.2
 * Fixed regression that didn't allow retaining walls and quays to be removed.
 * Improved replace tool mode's ability to swap configurations. 
+* Improved forced ground mode from an elevated/lowered node.
+* Updated Localization